laravel 8 update' :

+++ kbs2.1/app/Http/Controllers/FaqController.php
@@ -2,10 +2,9 @@
 
 namespace App\Http\Controllers;
 
-use App\Services\FaqService;
-use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
-use Illuminate\Support\Facades\Auth;
+use App\Services\FaqService;
+use Auth;
 
 class FaqController extends Controller
 {
@@ -20,18 +19,16 @@ class FaqController extends Controller
 
     public function index(Request $request)
     {
-        return $this->faqService->paginateData($request);
-//        if(Auth::user()->can('faq-list')) {
-//
-//            return $this->faqService->paginateData($request);
-//
-//        } else {
-//
-//            return response()->json(['status_code' => 424, 'messages'=>'User does not have the right permissions']);
-//
-//        }
+        
+        if(Auth::user()->can('faq-list')) {
 
-        // return $this->faqService->paginateData($request);
+            return $this->faqService->paginateData($request);
+
+        } else {
+
+            return response()->json(['status_code' => 424, 'messages'=>'User does not have the right permissions']);
+
+        }
 
     }
 
@@ -44,48 +41,28 @@ class FaqController extends Controller
 
 
     /**
-     * @return void
-     */
-    public function create()
-    {
-
-        /*return view('categorys.create',compact('permissions'));*/
-
-    }
-
-
-    /**
      * @param Request $request
      * @return JsonResponse
      */
     public function store(Request $request)
     {
-        return $this->faqService->createItem($request);
+        
+        if(Auth::user()->can('faq-create')) {
 
-//        if(Auth::user()->can('faq-create')) {
-//
-//            return $this->faqService->createItem($request);
-//
-//        } else {
-//
-//            return response()->json(['status_code' => 424, 'messages'=>'User does not have the right permissions']);
-//        }
+            return $this->faqService->createItem($request);
+
+        } else {
+
+            return response()->json(['status_code' => 424, 'messages'=>'User does not have the right permissions']);
+        }
 
     }
 
-    /**
-     * @param $id
-     * @return void
-     */
-    public function edit($id)
-    {
-        //
-    }
-
-
+    
     public function show($id)
     {
-        /* if(Auth::user()->can('faq-list')) {
+
+        if(Auth::user()->can('faq-list')) {
 
             return $this->faqService->getById($id);
 
@@ -93,40 +70,38 @@ class FaqController extends Controller
 
             return response()->json(['status_code' => 424, 'messages'=>'User does not have the right permissions']);
 
-        } */
-
-        return $this->faqService->getById($id);
+        }
 
     }
 
     public function update(Request $request)
     {
-        return $this->faqService->updateItem($request);
-//        if(Auth::user()->can('faq-edit')) {
-//
-//            return $this->faqService->updateItem($request);
-//
-//        } else {
-//
-//            return response()->json(['status_code' => 424, 'messages'=>'User does not have the right permissions']);
-//
-//        }
+        
+        if(Auth::user()->can('faq-edit')) {
+
+            return $this->faqService->updateItem($request);
+
+        } else {
+
+            return response()->json(['status_code' => 424, 'messages'=>'User does not have the right permissions']);
+
+        }
 
     }
 
 
     public function destroy(Request $request)
     {
-        return  $this->faqService->deleteItem($request->id);
-//        if(Auth::user()->can('faq-delete')) {
-//
-//            return  $this->faqService->deleteItem($request->id);
-//
-//        } else {
-//
-//            return response()->json(['status_code' => 424, 'messages'=>'User does not have the right permissions']);
-//
-//        }
+        
+        if(Auth::user()->can('faq-delete')) {
+
+            return  $this->faqService->deleteItem($request->id);
+
+        } else {
+
+            return response()->json(['status_code' => 424, 'messages'=>'User does not have the right permissions']);
+
+        }
 
     }
 }
